requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
prisma>=0.15.0
asyncpg>=0.29.0
openai>=1.0.0
python-multipart>=0.0.6
bcrypt>=4.0.0
pyjwt>=2.8.0
